hpdcache: add support for virtually-indexed addressing
This addition allows to send the target address in two steps:

  1. The first step contains the index of the cache (set and byte offset)
  2. The second step contains the tag.

In case of memory subsystems with virtual addressing, these two steps
enable the translation of the address by a Translation Lookaside Buffer
(TLB), in parallel of the indexing of the cache. This is what it is
called virtually-indexed physically-tagged addressing.

This modification adds additional signals in the interface of the
HPDcache:

  - The address is split in two fields (index and tag).
  - An additional tag signal is added per requester that contains the
    valid tag one cycle after the request (with the cache index) is
    accepted.
  - An abort signal is also added that allows the requester to abort a
    request transmitted the previous cycle.