ctrl: add support for the random victim selection policy
This commit adds support for the (pseudo) random victim selection
policy. The user can choose among: random or PLRU.

To choose, there is a new configuration variable in the parameters
package: PARAM_VICTIM_SEL. It can be set to:

    - 0: RANDOM
    - 1: PLRU

The random selection policy uses a 8-bit LFSR to generate a pseudo
random sequence of ways when replacing a cacheline.