Correctly handle write miss with error from memory
When there is a write miss in write-back mode, the cache issues a read
request to the memory on the missing cacheline, and put on-hold the
write while waiting for the refill.

If the refill is an error, the write request in the RTAB is tagged with
a bit error, and when replayed, it is immediately responded to the core
with the error flag set.