Add lowLatency parameter (#69)
The lowLatency parameter allows to configure the latency of loads and stores in the HPDcache.

When lowLatency parameter is set to 1, the HPDcache behaves as before this PR: Latency for stores and load hit is 1 cycle. This mode provides the lowest access latency but may exhibit lower throughput because of some structural hazards between loads and stores on the data SRAMs.

When lowLatency parameter is set to 0, the latency for stores and load hits is 2 cycles. This mode provides higher throughput (but also higher latency) because some structural hazards are prevented.

---------

Co-authored-by: Arnau Bigas <arnau.bigas@bsc.es>