ctrl: fix issue on write miss with the coalescing buffer full
The old implementation put the request in the RTAB with a wrong dependency
bit. There was not appropiate bit currently implemented for this case.

Instead of adding an additional dependency bit to the RTAB, this commit
does the following for that case:

- If the coalescing buffer is full, the cache triggers the read miss
  (because of the write allocate policy), but puts the write into the
  RTAB (which waits for the corresponding read miss to complete).

- If the coalescing buffer is not full, the cache triggers the read miss
  and put the write into the coalescing buffer. No entry of the RTAB is
  used in this case.

This has the effect that the CBUF becomes an extension of the RTAB. When
there are CBUFs available for a write miss, those are used. Otherwise,
RTAB entries are used. This means that in practice, there can be up to
CBUF entries + RTAB entries pending write misses.