uc: fix reset of req buffer control signals
To save some area, the design does not reset all FFs. We choose to reset
only FFs that are implicated in the control path.

In the Uncached/AMO handler (UC) there were some control FFs that were
not reset. This commit fixes this.

Previous version to this commit was functional as those registers were
only read under some specific circumstances. However this fix should
improve power consumption and spurious assertions in simulation when the
design is not fully reset yet.